% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2_interact3.R
\name{cov2_interact3}
\alias{cov2_interact3}
\title{Extract regression model data for 3 way interaction
with two covariates}
\usage{
cov2_interact3(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value
}
\description{
Pull the regression coefficient, standard error, t-value and p-value
for the three way interaction specified in your linear model.
}
\examples{
example <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width * Sepal.Width, data = iris)
cov2_interact3(example)
}

