% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_interact12.R
\name{cov_interact12}
\alias{cov_interact12}
\title{Extract regression model data for 2 way interaction (predictor 1 & 2)
with one covariate}
\usage{
cov_interact12(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value
}
\description{
Pull the regression coefficient, standard error, t-value and p-value
for the two way interaction between the first and second predictors
specified in your linear model. If you have no covariates in your model,
use interact12 (for 2 covariates, use cov2_interact12).
}
\examples{
example <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width * Sepal.Width, data = iris)
interact12(example)
}

