% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{covariate}
\alias{covariate}
\title{Extract regression model data for covariate 1}
\usage{
covariate(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value
}
\description{
Pull the regression coefficient, standard error, t-value and p-value
for the first covariate specified in your linear model.
}
\examples{
example <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width * Sepal.Width, data = iris)
covariate(example)
}

