% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate2.R
\name{covariate2}
\alias{covariate2}
\title{Extract regression model data for covariate 2}
\usage{
covariate2(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value
}
\description{
Pull the regression coefficient, standard error, t-value and p-value
for the second covariate specified in your linear model.
}
\examples{
example <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width * Sepal.Width, data = iris)
covariate2(example)
}

