% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact23.R
\name{interact23}
\alias{interact23}
\title{Extract regression model data for 2 way interaction (predictor 2 & 3)}
\usage{
interact23(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value
}
\description{
Pull the regression coefficient, standard error, t-value and p-value
for the two way interaction between the first and third predictors
specified in your linear model. If you have a covariate in your model,
use cov_interact23 (for 2 covariates, use cov2_interact23).
}
\examples{
example <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width * Sepal.Width, data = iris)
interact23(example)
}

