% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor1.R
\name{predictor1}
\alias{predictor1}
\title{Extract regression model data for predictor 1}
\usage{
predictor1(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value
}
\description{
Pull the regression coefficient, standard error, t-value and p-value
for the first predictor specified in your linear model.
}
\examples{
model <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width, data = Study1)
predictor1(model)
}

