% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor2.R
\name{predictor2}
\alias{predictor2}
\title{Extract regression model data for predictor 2}
\usage{
predictor2(lm)
}
\arguments{
\item{lm}{linear model}
}
\value{
beta, se, t-value, p-value
}
\description{
Pull the regression coefficient, standard error, t-value and p-value
for the second predictor specified in your linear model.
}
\examples{
example <- lm(formula = Sepal.Length ~ Petal.Length * Petal.Width, data = iris)
predictor2(example)
}

